# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product Catalog External API",
        "description": "API externe pour l'intégration du catalogue produits dans d'autres outils.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/ext",
            "description": "API Externe"
        }
    ],
    "paths": {
...
        "/manufacturerscolors": {
            "get": {
                "tags": [
                    "Couleurs"
                ],
                "summary": "Liste des couleurs fabricants",
                "description": "Retourne la liste des couleurs fabricants avec leur couleur principale associée",
                "operationId": "getManufacturerColors",
                "parameters": [
                    {
                        "name": "manufacturer",
                        "in": "query",
                        "description": "Filtrer par UUID de fabricant",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "primary_color",
                        "in": "query",
                        "description": "Filtrer par UUID de couleur principale",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des couleurs fabricants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "uuid": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "hex_code": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "color_sku_code": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "rgb": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "pantone_c": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "pantone_tcx": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "pms": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "primary_color": {
                                                        "properties": {
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "manufacturer": {
                                                        "properties": {
                                                            "uuid": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}